requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=0.1=cpu_0
asgiref=3.3.4=pyhd8ed1ab_0
astroid=2.2.5=pypi_0
autopep8=1.4.4=pypi_0
blas=1.0=mkl
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h294d835_1001
bzip2=1.0.8=he774522_0
ca-certificates=2021.5.30=h5b45459_0
certifi=2019.6.16=pypi_0
cffi=1.14.5=py38hcd4344a_0
chardet=3.0.4=pypi_0
charls=2.2.0=h6c2663c_0
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38he774522_0
dask-core=2021.5.0=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.6.0=pypi_0
dj-database-url=0.5.0=py_0
django=2.2.13=pypi_0
django-allauth=0.39.1=pypi_0
django-cors-headers=3.0.2=pypi_0
django-heroku=0.3.1=pyh8c360ce_0
django-rest-auth=0.9.5=pypi_0
djangorestframework=3.9.4=pypi_0
filelock=3.0.12=pyh9f0ad1d_0
freetype=2.10.4=hd328e21_0
fsspec=0.9.0=pyhd3eb1b0_0
gdown=3.13.0=pyhd8ed1ab_0
giflib=5.2.1=h62dcd97_0
gunicorn=19.9.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.8=pypi_0
imagecodecs=2021.4.28=py38h5da4933_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2019.4=245
isort=4.3.21=pypi_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py38hd77b12b_0
krb5=1.19.1=hbae68bd_0
lazy-object-proxy=1.4.1=pypi_0
lcms2=2.12=h83e58a3_0
lerc=2.2.1=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.7=h2bbff1b_5
libmklml=2019.0.5=haa95532_0
libpng=1.6.37=h2a8f88b_0
libpq=13.3=hfcc5ef8_0
libtiff=4.2.0=hd0e1b90_0
libzopfli=1.0.3=ha925a31_0
locket=0.2.1=py38haa95532_1
lz4-c=1.9.3=h2bbff1b_0
matplotlib-base=3.3.4=py38h49ac443_0
mccabe=0.6.1=pypi_0
mkl=2019.4=245
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.0=py38hf9181ef_0
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=h6d14046_1
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.0.1=pypi_0
olefile=0.46=py_0
openjpeg=2.3.0=h5ec785f_1
openssl=1.1.1k=h8ffe710_0
partd=1.2.0=pyhd3eb1b0_0
pep8=1.7.1=pypi_0
pillow=8.2.0=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
psycopg2=2.8.6=py38hd8c33c5_3
pycodestyle=2.5.0=pypi_0
pycparser=2.20=py_2
pylint=2.3.1=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py38haa244fe_3
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.1=pyhd3eb1b0_0
python3-openid=3.1.0=pypi_0
python_abi=3.8=1_cp38
pytorch=1.6.0=cpu_py38h538a6d7_0
pytz=2019.1=pypi_0
pywavelets=1.1.1=py38he774522_2
pyyaml=5.4.1=py38h2bbff1b_1
requests=2.22.0=pypi_0
requests-oauthlib=1.2.0=pypi_0
scikit-image=0.18.1=py38hf11a4ad_0
scipy=1.6.2=py38h14eb087_0
setuptools=52.0.0=py38haa95532_0
six=1.12.0=pypi_0
snappy=1.1.8=h33f27b4_0
sqlite=3.35.4=h2bbff1b_0
sqlparse=0.3.0=pypi_0
tifffile=2021.4.8=pyhd3eb1b0_2
tk=8.6.10=he774522_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.7.0=py38_cu102
tornado=6.1=py38h2bbff1b_0
tqdm=4.61.0=pyhd8ed1ab_0
typed-ast=1.4.0=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.25.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
whitenoise=5.2.0=py_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
wrapt=1.11.2=pypi_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
gunicorn==20.1.0